commit 3 "couche web"

+++ src/main/java/com/mnr/gestioncinema/dao/CategorieRepository.java
@@ -9,5 +9,5 @@ import java.math.BigDecimal;
 
 //add api rest
 @RepositoryRestResource
-public interface CategorieRepository extends JpaRepository<Categorie, BigDecimal> {
+public interface CategorieRepository extends JpaRepository<Categorie, Long> {
 }
