+++ src/main/java/com/mnr/gestioncinema/dao/CinemaRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import java.math.BigDecimal;
 
 @RepositoryRestResource
-public interface CinemaRepository extends JpaRepository<Cinema, BigDecimal> {
+public interface CinemaRepository extends JpaRepository<Cinema, Long> {
 }
