+++ src/main/java/com/mnr/gestioncinema/dao/FilmRepository.java
@@ -8,5 +8,5 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import java.math.BigDecimal;
 
 @RepositoryRestResource
-public interface FilmRepository extends JpaRepository<Film, BigDecimal> {
+public interface FilmRepository extends JpaRepository<Film, Long> {
 }
